Switch dispersion to double input for now

+++ spotfind_bm.cxx
@@ -61,7 +61,7 @@ static void BM_extended_dispersion_gain(benchmark::State& state) {
 BENCHMARK_TEMPLATE(BM_extended_dispersion_gain, double)->Unit(benchmark::kMillisecond);
 
 static void BM_ISPC(benchmark::State& state) {
-  ImageSource<float, float> src;
+  ImageSource<double, float> src;
 
   // Convert mask, dst to int because of https://github.com/ispc/ispc/issues/1709
   std::unique_ptr<int[]> mask(new int[IMAGE_W * IMAGE_H]);
